Change the file structure
