--- src/main/java/se/swcg/consultauction/client/controller/ClientController.java
@@ -1,4 +1,0 @@
-package se.swcg.consultauction.client.controller;
-
-public class ClientController {
-}
