--- src/main/java/se/swcg/consultauction/client/dto/ClientDTO.java
@@ -1,4 +1,0 @@
-package se.swcg.consultauction.client.dto;
-
-public class ClientDTO {
-}
