--- src/main/java/se/swcg/consultauction/client/repository/ClientRepository.java
@@ -1,4 +1,0 @@
-package se.swcg.consultauction.client.repository;
-
-public class ClientRepository {
-}
