--- src/main/java/se/swcg/consultauction/client/service/ClientService.java
@@ -1,4 +1,0 @@
-package se.swcg.consultauction.client.service;
-
-public interface ClientService {
-}
