--- src/main/java/se/swcg/consultauction/client/service/ClientServiceImpl.java
@@ -1,4 +1,0 @@
-package se.swcg.consultauction.client.service;
-
-public class ClientServiceImpl {
-}
