+++ src/main/java/se/swcg/consultauction/controller/ClientController.java
@@ -1,0 +1,4 @@
+package se.swcg.consultauction.controller;
+
+public class ClientController {
+}
