+++ src/main/java/se/swcg/consultauction/dto/ClientDTO.java
@@ -1,0 +1,4 @@
+package se.swcg.consultauction.dto;
+
+public class ClientDTO {
+}
