+++ src/main/java/se/swcg/consultauction/repository/ClientRepository.java
@@ -1,0 +1,4 @@
+package se.swcg.consultauction.repository;
+
+public class ClientRepository {
+}
