+++ src/main/java/se/swcg/consultauction/service/ClientService.java
@@ -1,0 +1,4 @@
+package se.swcg.consultauction.service;
+
+public interface ClientService {
+}
